contracts: Refactor the contracts call runtime API
