+++ frame/contracts/common/Cargo.toml
@@ -14,6 +14,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 # This crate should not rely on any of the frame primitives.
+bitflags = "1.0"
 codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false, features = ["derive"] }
 sp-std = { version = "2.0.0", default-features = false, path = "../../../primitives/std" }
 sp-runtime = { version = "2.0.0", default-features = false, path = "../../../primitives/runtime" }
