+++ frame/contracts/common/src/lib.rs
@@ -18,13 +18,28 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+use bitflags::bitflags;
+use codec::{Decode, Encode};
+use sp_runtime::{DispatchError, RuntimeDebug};
 use sp_std::prelude::*;
 
-/// A result type of a get storage call.
+/// A result type of a `bare_call` call.
+/// The result of a contract exection together together with the consumed gas.
+#[derive(Eq, PartialEq, Encode, Decode, RuntimeDebug)]
+pub struct ContractExecResult {
+	pub exec_result: ExecResult,
+	pub gas_consumed: u64,
+}
+
+/// A result type of a `get_storage` call.
 pub type GetStorageResult = Result<Option<Vec<u8>>, ContractAccessError>;
 
+/// A result type of a `rent_projection` call.
+pub type RentProjectionResult<BlockNumber> =
+	Result<RentProjection<BlockNumber>, ContractAccessError>;
+
 /// The possible errors that can happen querying the storage of a contract.
-#[derive(Eq, PartialEq, codec::Encode, codec::Decode, sp_runtime::RuntimeDebug)]
+#[derive(Eq, PartialEq, Encode, Decode, RuntimeDebug)]
 pub enum ContractAccessError {
 	/// The given address doesn't point to a contract.
 	DoesntExist,
@@ -32,11 +47,7 @@ pub enum ContractAccessError {
 	IsTombstone,
 }
 
-/// A result type of a `rent_projection` call.
-pub type RentProjectionResult<BlockNumber> =
-	Result<RentProjection<BlockNumber>, ContractAccessError>;
-
-#[derive(Eq, PartialEq, codec::Encode, codec::Decode, sp_runtime::RuntimeDebug)]
+#[derive(Eq, PartialEq, Encode, Decode, RuntimeDebug)]
 pub enum RentProjection<BlockNumber> {
 	/// Eviction is projected to happen at the specified block number.
 	EvictionAt(BlockNumber),
@@ -45,3 +56,62 @@ pub enum RentProjection<BlockNumber> {
 	/// E.g. because the contract accumulated enough funds to offset the rent storage costs.
 	NoEviction,
 }
+
+bitflags! {
+	/// Flags used by a contract to customize exit behaviour.
+	#[derive(Encode, Decode)]
+	pub struct ReturnFlags: u32 {
+		/// If this bit is set all changes made by the contract exection are rolled back.
+		const REVERT = 0x0000_0001;
+	}
+}
+
+/// Output of a contract call or instantiation which ran to completion.
+#[derive(PartialEq, Eq, Encode, Decode, RuntimeDebug)]
+pub struct ExecReturnValue {
+	/// Flags passed along by `seal_return`. Empty when `seal_return` was never called.
+	pub flags: ReturnFlags,
+	/// Buffer passed along by `seal_return`. Empty when `seal_return` was never called.
+	pub data: Vec<u8>,
+}
+
+impl ExecReturnValue {
+	/// We understand the absense of a revert flag as success.
+	pub fn is_success(&self) -> bool {
+		!self.flags.contains(ReturnFlags::REVERT)
+	}
+}
+
+/// Call or instantiate both call into other contracts and pass through errors happening
+/// in those to the caller. This enum is for  the caller to distinguish whether the error
+/// happened during the execution of the callee or in the current execution context.
+#[derive(PartialEq, Eq, Encode, Decode, RuntimeDebug)]
+pub enum ErrorOrigin {
+	/// The error happened in the current exeuction context rather than in the one
+	/// of the contract that is called into.
+	Caller,
+	/// The error happened during execution of the called contract.
+	Callee,
+}
+
+/// Error returned by contract exection.
+#[derive(PartialEq, Eq, Encode, Decode, RuntimeDebug)]
+pub struct ExecError {
+	/// The reason why the execution failed.
+	pub error: DispatchError,
+	/// Origin of the error.
+	pub origin: ErrorOrigin,
+}
+
+impl<T: Into<DispatchError>> From<T> for ExecError {
+	fn from(error: T) -> Self {
+		Self {
+			error: error.into(),
+			origin: ErrorOrigin::Caller,
+		}
+	}
+}
+
+/// The result that is returned from contract execution. It either contains the output
+/// buffer or an error describing the reason for failure.
+pub type ExecResult = Result<ExecReturnValue, ExecError>;
