+++ frame/contracts/src/benchmarking/mod.rs
@@ -34,6 +34,7 @@ use frame_system::{Module as System, RawOrigin};
 use parity_wasm::elements::{Instruction, ValueType, BlockType};
 use sp_runtime::traits::{Hash, Bounded};
 use sp_std::{default::Default, convert::{TryInto}};
+use pallet_contracts_primitives::RentProjection;
 
 /// How many batches we do per API benchmark.
 const API_BENCHMARK_BATCHES: u32 = 20;
