+++ frame/contracts/src/gas.rs
@@ -14,12 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with Substrate. If not, see <http://www.gnu.org/licenses/>.
 
-use crate::{Trait, exec::ExecError};
+use crate::Trait;
 use sp_std::marker::PhantomData;
 use sp_runtime::traits::Zero;
 use frame_support::dispatch::{
 	DispatchResultWithPostInfo, PostDispatchInfo, DispatchErrorWithPostInfo,
 };
+use pallet_contracts_primitives::ExecError;
 
 #[cfg(test)]
 use std::{any::Any, fmt::Debug};
